refactor: NumberGenerator 인터페이스와 구현체 리팩토링
NumberGenerator 인터페이스와 구현체 리팩토링
