+++ src/main/java/baseball/numbergenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.numbergenerator;
 
 public interface NumberGenerator {
     public Integer generate();
