+++ src/main/java/baseball/numbergenerator/NumberGeneratorSequenceImpl.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.numbergenerator;
 
 import java.util.List;
